requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Metrics and evaluation
tqdm>=4.65.0

# Data augmentation for contrastive learning
nlpaug>=1.1.11

# Sentence transformers (useful for TSDAE and SimCSE baselines)
sentence-transformers>=2.2.0

# Utilities
pyyaml>=6.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: for advanced augmentations
# textaugment>=1.3.4
# backtranslation>=0.0.1
